Define some wallpaper groups in python
Defining the wallpaper groups is a much simpler and clearer operation in python,
which is why I have implemented the definition of the wallpaper groups
in python. The key factor for this is being able to define keyword
arguments, which means all the values being passed to the constructor
are labelled.

This also defines a subset of all the wallpaper groups which allows for
testing the overall functionality before I work on copying all the
values over.