Get package to install properly with poetry
This gets the poetry install to work correctly ensuring all the required
components are in the correct place and installed properly. The
structure of the project is such that there is the packing module, which
is the C++ portion. This is then compiled to the _packing module for
importing into the pypacking module, which is the user interface
component.