STY Use const wherever possible, init with {}
This adds the use of const wherever possible, which restricts the
modification of variables far more. This is a more defensive strategy
than not having the consts there.

An additional style change, which is somewhat inherited from the move to
const variables is the initialisation of variables with the {} syntax.