ENH Subclass Shape in python for interesting shapes
Rather than attempting to create all the shapes statically in C++ I am
creating the shapes in Python, which has the added benefit of being able
to create shapes in a far more interactive manner.

Additionally the tools for creating the arrays of points are much
simpler to use in python, particularly with numpy, making this the
preferred language for building these shapes.