ENH Export WallpaperGroup to python for testing
This exports the WallpaperGroup class, in addition to all the component
classes, enabling the testing of the class in python. This will also
allow for python to create these classes dynamically.

As part of the export process I have had to modify and create a series
of constructors for the classes. These classes are also all containing
const elements, since they are not supposed to change after startup, and
it is easier to prohibit them from changing rather than checking they
don't change.

Since the WallpaperGroup classes are effectively just holding some data,
I want to be able to populate all the instances of the classes
dynamically using python. Taking a data file from an appropriate source,
being a Crystallographic database, parsing the file and generating all
the Wallpaper Groups.

For 3D crystals there are too many combinations to enter manually, so
being able to generate everything from a datafile is going to be key to
getting everything working.