ENH Properly export classes in wallpaper module
This implements all the functions for inclusion in the wallpaper module,
including adding it to the _packing module. There were a number of
methods which had been left unimplemented which are now implemented,
including copy constructors, which were not created by default due to
using const values.

The lack of a default copy constructor probably indicates I should be
using a shared pointer rather than creating new instances of the class,
although that is an optimisation for another commit.